Add files via upload
In this project ill be working on a tour project which will render the list of all available touring locations. In this project ill make use of the useEffect hook in react to fetch data from an api once the screen is rerendered. Ill use useState to store states involved and many more